ST Port: CCR use 16 bit shorts so the overlap tests are word compares.

+++ tiberiandawn/aircraft.cpp
@@ -944,7 +944,7 @@ short const* AircraftClass::Overlap_List(void) const
     return (Class->Overlap_List());
 }
 
-void AircraftClass::Get_AABB(int& x0, int& y0, int& x1, int& y1) const
+void AircraftClass::Get_AABB(short& x0, short& y0, short& x1, short& y1) const
 {
     int const x = Coord_X(Coord);
     int const y = Coord_Y(Coord);
